github_populator/requirements.txt
requests>=2.25.1
PyGithub>=1.55
python-gitlab>=3.0.0
gitpython>=3.1.0
astor>=0.8.1
black>=21.9b0
flake8>=3.9.0
pytest>=6.2.4
python-dotenv>=0.19.0
google-cloud-aiplatform>=1.38.0